Update access when deleting project collaborator

+++ backend/app/api/routes/projects.py
@@ -2032,7 +2032,7 @@ def put_project_collaborator(
         raise HTTPException(404, "User not found")
     token = users.get_github_token(session=session, user=current_user)
     url = (
-        f"https://api.github.com/repos/{owner_name}/{project_name}/"
+        f"https://api.github.com/repos/{project.github_repo}/"
         f"collaborators/{github_username}"
     )
     resp = requests.put(url, headers={"Authorization": f"Bearer {token}"})
@@ -2068,16 +2068,27 @@ def delete_project_collaborator(
     current_user: CurrentUser,
     session: SessionDep,
 ) -> Message:
-    app.projects.get_project(
+    project = app.projects.get_project(
         owner_name=owner_name,
         project_name=project_name,
         session=session,
         current_user=current_user,
         min_access_level="admin",
     )
+    user = session.exec(
+        select(User)
+        .join(User.account)
+        .where(Account.github_name == github_username)
+    ).first()
+    logger.info(
+        f"Fetched user account {user.email} with GitHub username "
+        f"{github_username}"
+    )
+    if user is None:
+        raise HTTPException(404, "User not found")
     token = users.get_github_token(session=session, user=current_user)
     url = (
-        f"https://api.github.com/repos/{owner_name}/{project_name}/"
+        f"https://api.github.com/repos/{project.github_repo}/"
         f"collaborators/{github_username}"
     )
     resp = requests.delete(url, headers={"Authorization": f"Bearer {token}"})
@@ -2086,6 +2097,20 @@ def delete_project_collaborator(
             f"Failed to delete collaborator ({resp.status_code}): {resp.text}"
         )
         raise HTTPException(resp.status_code)
+    access = session.exec(
+        select(UserProjectAccess)
+        .where(UserProjectAccess.user_id == user.id)
+        .where(UserProjectAccess.project_id == project.id)
+    ).first()
+    if access is not None:
+        access.access = None
+    else:
+        session.add(
+            UserProjectAccess(
+                user_id=user.id, project_id=project.id, access=None
+            )
+        )
+    session.commit()
     return Message(message="Success")
